Rewrite json_converter to function as a reformatter
With this change, we now use json_formatter.c to reformat JSON instead
of JS functions. This might be faster; more importantly, by not parsing
we keep the exact same formatting of numbers, strings, etc.

Fixed some small bugs involving what happens when OOM, and expanded the
memory size to 8MB to allow for the larger size allocated by the
formatter.